w3c-reports/12092014/efl/runned-tests/html/webappapis/scripting/processing-model-2/compile-error-cross-origin-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token ')'

PASS window.onerror - compile error in <script src=//www1...> 
PASS window.onerror - compile error in <script src=//www1...> (column) 
